// web/src/API.ts
export type Note = {
  __typename: "Note",
  id?: string | null,
  title?: string | null,
  content?: string | null,
  createdAt: string,
  updatedAt: string,
};